chore(rover): refactor to enable testing, add tests
